Updated UCS script
Added option to choose all controllers and removed to forcing all controllers, to enable texture detection for the X360 (different layout)
Used the tools subfolder for the exe files for a cleaner looking archive
Fixed a bug with the info display
Fixed a bug when no entry was given when manually selecting a button prompt flash colour

Alchemy:
Added a code (unused) which I used to list animations based on fightmove names (for fightstyle_default.xmlb). It's a neat little Powershell code.