Fixed a bug with movies
ui/menus/movie.xmlb is no longer deleted and there don't seem to be any more crashes.
I also kept the package which matches with the xmlb file (but it doesn't seem to be required like so many package files).
The files use x2m_movies.igb, but it seems to work without the file. If there are issues, we'll have to change the removal of x2m files.